add attributes to Authentication Reply Packets in scuba
Summary:
A fair number of users and Netsec have asked for the capability to have a single way to filter scuba records through a user -> AAA status mapping. Network Threat detection team depends on such a mapping to perform data analysis and anomaly detection on TACACS records. This was hard to do in the original traffic logging, since tacquito's logger logged TACACS traffic, which does not contain all attributes in the Reply packets.
This diff will add attributes into Authen Reply packets so that users can now filter these records easily in scuba.
I have added a benchmark for the new contextual logger, so we can check for performance gains/loss in case the Gather method is modified.

Reviewed By: nprintz

Differential Revision: D40878839

fbshipit-source-id: 765a83d1451131f97b04882765b3a017cf55563a