add attributes to Accounting Reply packets
Summary: Because start.go attaches an ResponseLogger to the incoming request, that logger needs to be overwritten with new context, which is why I had to introduce a new PopWriter method which would remove the last Writer from the list of response writers. We can then attach a new writer with updated context to the response logger. The pop writer method would be required by authorization and accounting packets

Reviewed By: nprintz

Differential Revision: D43142265

fbshipit-source-id: 02b666392dc712972ab0da5390b71e1d375e7bec