add a context logger
Summary:
This diff is part of a stack that aims to simplify contextual logging within tacquito

Add a Set method to the existing logger. This method save values into a context, and then return the context. This context can be used by callsites to set other contexts (such as the response context: see bottom diff).

Reviewed By: nprintz

Differential Revision: D44072481

fbshipit-source-id: 7fec7ab671f94d790cf7e34dd8195922244deb3b