add a Context function to response
Summary:
This diff is part of a stack that aims to simplify contextual logging within tacquito

tacquito's response struct allows callsites to specify additional writers. This comes in handy in situations when the server wants to send output to a place other than the original connection object

These writers conform to the io.Writer interface, but this interface is restrictive because it doesn't allow the callsites to utilize any context values. Context is the major way we store values within tacquito that need to persist. This diff
1) Introduce a new Writer interface with ctx allowed as an argument to the Write function
1) Introduces a new Context(ctx context.Context) function to the Response interface which will set call sites pass in a ctx to replace the original response context. There is no need to get the context of the response right now, so there is no API as of yet to pull the ctx value from the response.
2) Make the response logger's writer conform to the new Writer interface

Reviewed By: nprintz

Differential Revision: D44072482

fbshipit-source-id: 1a356e9afc6315018cf02561b29bcfd6035e8206