add a scoped logger
Summary:
This diff adds a handler scoped logger, which can store values that should persist across the lifetime of a handler. The values are stored into a handler's context, which is initialized to the value of the context embedded into the request at the start of the handler state machine.
The logger conforms to a private *recorder* interface, which opens up the possibility to inject this dependency into the handlers in the future

Reviewed By: nprintz

Differential Revision: D44072483

fbshipit-source-id: 66a277e6ec04d604488c03c2216720d3cc8ae8a6