Move heavy lifting into goroutine so accept loop frees up
Summary: As part of Tacquito's perf optimization, I want to make sure that the work being done in the accept loop is minimal (which is the general pattern in most TCP servers). For this I would need to separate out the secret get logic into a goroutine. This diff on it's own doesn't have a huge impact on rps directly, but is useful when we start using a worker pool later

Reviewed By: nprintz

Differential Revision: D45550985

fbshipit-source-id: 5beee3b716702f40f7563fdea03fd17bc05d7f9f