dont validate args length if its accounting packet
Summary:
we see that some accounting packets in ciena 6500 and ZPE devices are sending empty args of length 0 as below:
{F1021587052}

the RFC says the args need to be between length 2 to 255. Moreover there is no value in sending empty args to the server.

But since this is easier fixed in the server than by the vendor, making this change.

Reviewed By: RollerMatic

Differential Revision: D46547099

fbshipit-source-id: 191b4a44fb7ca0b85d9586d3b80404a07b76e228