unblock loader when context is cancelled
Summary:
BlockUntilLoaded is supposed to prevent the server from serving request till the config is loaded. This has an unintended side effect of preventing the process from existing from SIGTERM if for some reason the main.go's context is cancelled.
In my case, I accidentally started two instances of the server . Since the thrift port is hardcoded to 2047 for local instances, thrift binding failed which led to cancellation of the context before config could be subscribed. This led to the main goroutine in a limbo state.

I am adding a fix where the BlockUntilLoaded function watches for context cancellation in addition to config loading

Reviewed By: shringiarpit26

Differential Revision: D47113946

fbshipit-source-id: 6fb4042dd152b273fa4bf8f9bc01c04070c34043