store the local address to know how many connection we get on primary/secondary vip
Summary:
I want to experiment for a tacprod like setup for tacquito, and have different TW tasks that advertise *either* the primary or the secondary VIP instead of both. The main reason is that currently if a client has trouble talking to the primary VIP (due to routing issues/server being overloaded), and tries to fallback to the secondary VIP then the underlay routing could take the request to the same server, which could make the load even worse.
I also want to explore if this helps in reducing the number of TCP RSTs that are being sent to devices. These RSTs manifest in the form of a error "unexpected EOF" on stderr.
But before we get on that train, I want to check how much traffic the VIPs get individually, and if the same task gets traffic destined for primary and secondary vips from the same client. This could be a pointer towards retries routing to the same server.

You might notice that the conn-local-address appears in the go's *unstripped* format [<ipv6>]:<port>. This is because I didn't want to spend additional CPU cycles (it's a per request cost) for formatting this IP because the formatting can be achieved in scuba as well.

Reviewed By: nprintz

Differential Revision: D47279910

fbshipit-source-id: 4408bafe0ea77acd84650181c4cb72ba54e74094