replace port/ip and bracket split functions with a single call
Summary:
we perform two transformations on the remote address : remove port, and remove brackets. These transformations are done at different places, and the second transformation (remove brackets) is done once for request, and once for response. Instead we can apply it directly at the source in the server loop.
I replaced the individual functions with a call to net.SplitHostPort

Reviewed By: nprintz

Differential Revision: D47387028

fbshipit-source-id: f23fce7000ddd7be480477f9fb9217dd71a6687f