loader's Get function within a goroutines
Summary:
Within tacquito, the loader's Get function is used to map an incoming request's IP to a scope, which is how tacquito is able to determine the Authenticator, Authorizer and Accounter for a particular user trying to login to a specific device (that is mapped to the aforementioned scope). The public Get function inserts the request into the Loader's query channel, and can therefore be blocking. ngoyal mentioned that this could be a limiter to our server's throughput.
Running a goroutine to serve the Get request will reduce the coupling of request processing at a time when our delay comes from CPU scheduling.
However, if the loader Get call is slow because of RPC delays, then this method might cause more problems because the server will accept requests but won't be able to serve them in the 5s budget allotted to tacquito for AAA transactions i.e there will be less backpressure from loader to the Accept loop. Perhaps we could wrap our RPCs with timeouts

Reviewed By: shringiarpit26

Differential Revision: D47560028

fbshipit-source-id: 46c831dbf22a1f6c6cd8edb3f7449f60ad5863e1