cleaning up unused functions in Log interface for Secret providers
Summary:
the Record method is unnecessary inside the Providers, because the tacacs request isn't ready by the time the control logic comes to the provider Get function. Once the callsite receives the PSK from Provider's Get, the server can decrypt the TACACS packet bytes and *create* the TACACS request struct. Removing the function from the private interface `loggerProvider`

This diff also includes a small logging change from an Infof call -> Debugf call. The line is rarely used in debugging, and should not be sent to the scuba `stderr` table.

Reviewed By: elizabethjoshi

Differential Revision: D53206362

fbshipit-source-id: 007116d779d12d8bf2147ceef8141e6c038ecf44