stop sending the server message on successful Authorization responses
Summary:
One of the observations gathered from testing tacquito in corp is that Cisco Cat 9Ks echo the content of the Server message field inside the Authorization Response packet to the remote vty line, causing automation workflows which depend on screen scraping to break and also leads to a very noisy interaction with the client.

```
enr101-00-off-sw3#conf t
authorization approved <-------- unexpected

Enter configuration commands, one per line.  End with CNTL/Z.
enr101-00-off-sw3(config)#
```

I raised this to [CISCO](https://my.workplace.com/groups/112344676125670/permalink/1343560813004044/), and they don't have any knob that can turn this behaviour off.  Quoting the response from our internal POC
```
per the RFC, Cisco is still adhering to it as the section in question states:

server_msg, server_msg_len

This is a string that *may* be presented to the user. The server_msg_len indicates the length of the server_msg field, in bytes. For details of text encoding, see "Treatment of Text Strings" (Section 3.7).

if that may was a shouldn’t, the behavior observed on the Catalyst 9K would an RFC violation and would be treated as a bug.
```

If we need to continue onboarding corp to tacquito, we need to not send the ServerMsg string to downstream devices.

We could disable sending the server msg on a per scope basis (it would need some code changes, but can be done), or disable sending the ServerMsg attribute entirely on successful responses. The latter might seem a little absurd, but the initial reason why we had needed the ServerMsg attribute was to provide contextuals hints around AuthorStatusError/AuthorStatusFail messages to the downstream network device, and log this hint to the scuba table. For successful Authorization Responses, the ServerMsg field is repetitive since it provides [no additional information](https://fburl.com/scuba/tacquito/o9wn58y6) to the transaction. The Packet Response codes of AuthorStatusPassAdd/AuthorStatusPassRepl are enough in these scenarios. Since a successful authorization response is something that is more prevalent than denies, we are sending more bytes across the wire than needed, leading to some perf degradation. I think the problem is worse for accounting (https://fburl.com/scuba/tacquito/v1gsxu6r), and we should probably stop sending a server message on successful start/stop too.

TPH reports an approx 7% increase in rpm https://pxl.cl/4Jbfr  when comparing RPMs between prod tacquito, and a canaried task (running this code). Second peak is the RPM from the canaried task.

Reviewed By: nprintz

Differential Revision: D56271364

fbshipit-source-id: 1057ca1efbb4797f16d6a8afe5dcc9bbb69cb2cf