Send AuthenStatusFail on zero length password
Summary: Tacquito has two status codes to signal failure to devices AuthenStatusFail, and AuthenStatusError. All vendors within Prod and corp (that we have tested with) exhibit a fallback behaviour on AuthenStatusError i,e they fallback to local authentication (this is as per configuration). Tacquito's behaviour today is that an empty password triggers an AuthenStatusError, which could be an easy way for an attacker or a non malicious actor to trigger local authentication (and even local authorization post login for some vendors which don't use cmd based authorization) as long as they choose to send a <cr> on the password prompt. I am modifying this behaviour to instead shortcircuit to AuthenStatusFail in case the password field is empty.

Differential Revision: D58028453

fbshipit-source-id: 39b2a98ca6efc2255724605d3c18ed3cfb9b5384