Stop sending ServerMsg attribute on authen success
Summary: This was brought up during the corp onboarding process, that the authentication server msg gets echoed to the remote vty line , and bothers automation tooling. The "login success" server msg does not add much value to the actual status of the TACACS transaction. The Status field "AuthenStatusPass" already provides the same information. I am removing this field here to save some storage on the scuba table and  inconvenience on the client side.

Reviewed By: shringiarpit26

Differential Revision: D58029862

fbshipit-source-id: c82474b426a84f2a74671027ab5c97746499916b