remove unused UUID
Summary:
The UUID was initially meant to help in correlation between data in tacquito scuba, and data in tacquito-stderr scuba. Due to the recent logging changes, the data in both the scuba tables is mutually exclusive, so we don't need this additional logging anymore.

Once this diff lands, I can drop the column from stderr scuba

Reviewed By: nprintz

Differential Revision: D50895652

fbshipit-source-id: 2f023401f6feeb66b5cfc6d2e167a9d1319e5fbb