resolve additional matching issue (#11)
Summary:
Pull Request resolved: https://github.com/facebookincubator/tacquito/pull/11

Go's Match string function is similar to a substring matcher, as opposed to python's regex `match` function which matches from the beginning of the string. For example
`cat.*` will match the command `bash /etc/some_folder/file_name_contains_cat.sh`
as per go matchers.
An easy way to restrict this is to add boundary characters inside expressions as tacquito reads them.
There are two places we could add these boundary characters
1) config
2) the server code

Within facebook's infra , it would be okay to attach this to either of the places since we control the config. But the stringy module is a part of the OSS modules, and in that case it's better to add the guardrails to the server code so the boundaries are not *optional*

Reviewed By: elizabethjoshi

Differential Revision: D64148191

fbshipit-source-id: 16f45c8a3e21c4d5d0509d9fe20a52fdedcbc22c